+++ addons/mrp/i18n/mrp.pot
@@ -392,7 +392,7 @@ msgid "A Manufacturing Order is already done or cancelled."
 msgstr ""
 
 #. module: mrp
-#: code:addons/mrp/models/stock_warehouse.py:0
+#: code:addons/mrp/models/stock_orderpoint.py:0
 #, python-format
 msgid ""
 "A product with a kit-type bill of materials can not have a reordering rule."
